backend/requirements.txt
APScheduler==3.10.4
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
spacy==3.8.13
SQLAlchemy==2.0.49
